Куча изменений, зафиксировал рабочее состояние, чтобы сейчас чистить код

+++ src/userBot/bot.js
@@ -1,7 +1,7 @@
 const { Telegraf } = require('telegraf');
 const { MongoClient } = require('mongodb');
 const { session } = require('telegraf-session-mongodb');
-const StartCommand = require('./commands/startCommand');
+const StartCommand = require('../shared/commands/startCommand');
 const InfoCommand = require('./commands/infoCommand');
 const PortfolioCommand = require('./commands/portfolioCommand');
 const ContactCommand = require('./commands/contactCommand');
@@ -9,7 +9,8 @@ const PriceCommand = require('./commands/priceCommand');
 const CallbackHandler = require('./callbacks/callbackHandler');
 const ReminderService = require('./services/reminderService');
 const CleanUpService = require('./services/cleanUpService');
-const User = require('../db/models/user');
+const NotificationService = require('./services/notificationService');
+const User = require('../database/models/user');
 const Log = require('../lib/log');
 const ErrorHandler = require('../lib/errorHandler');
 const config = require('../config/config');
@@ -18,7 +19,7 @@ const cron = require('node-cron');
 
 class UserBot {
     constructor() {
-        this.bot = new Telegraf(config.telegramToken);
+        this.bot = new Telegraf(config.userBotToken);
         this.logger = new Log();
         this.mongoClient = null;
         ErrorHandler.setLogger(this.logger);
@@ -56,9 +57,7 @@ class UserBot {
      * Настраивает middleware для хранения сессий в MongoDB.
      */
     async setupSessionMiddleware() {
-        this.bot.use(
-            session(this.mongoClient.db(), { collectionName: 'sessions' })
-        );
+        this.bot.use(session(this.mongoClient.db(), { collectionName: 'sessions' }));
         this.logger.info('Middleware сессий установлен');
     }
 
@@ -66,13 +65,7 @@ class UserBot {
      * Регистрирует обработчики команд бота.
      */
     async registerCommandHandlers() {
-        const commands = [
-            StartCommand,
-            InfoCommand,
-            PortfolioCommand,
-            ContactCommand,
-            PriceCommand,
-        ];
+        const commands = [StartCommand, InfoCommand, PortfolioCommand, ContactCommand, PriceCommand];
 
         for (const Command of commands) {
             this.bot.command(Command.name, async (ctx) => {
@@ -106,17 +99,54 @@ class UserBot {
     async registerMessageHandler() {
         this.bot.on('message', async (ctx) => {
             const { from, message } = ctx;
-            this.logger.info(
-                `Получено сообщение от пользователя ${from.id}: ${message.text}`
-            );
-            await this.bot.telegram.sendMessage(
-                config.userId,
-                `Пользователь ${from.first_name} ${from.last_name} (@${from.username}) ${from.id} отправил сообщение: ${message.text}`
-            );
-            const messageUser = await ctx.reply(
-                'Ваше сообщение передано Администратору'
-            );
-            setTimeout(() => ctx.deleteMessage(messageUser.message_id), 3000);
+            this.logger.info(`Получено сообщение от пользователя ${from.id}`);
+
+            // Обработка текстового сообщения
+            if (message.text) {
+                this.logger.info(`Текст сообщения: ${message.text}`);
+                await this.bot.telegram.sendMessage(
+                    config.adminId,
+                    `Пользователь ${from.first_name} ${from.last_name} (@${from.username}) ${from.id} отправил сообщение: ${message.text}`
+                );
+                const messageUser = await ctx.reply('Ваше сообщение передано Администратору');
+                setTimeout(() => ctx.deleteMessage(messageUser.message_id), 3000);
+            }
+
+            // Обработка фотографии
+            if (message.photo) {
+                this.logger.info('Получено фотографическое сообщение');
+                const photo = await ctx.telegram.getFile(message.photo[message.photo.length - 1].file_id);
+                console.log(photo);
+                await this.bot.telegram.sendPhoto(config.adminId, photo.file_id, {
+                    caption: `Пользователь ${from.first_name} ${from.last_name} (@${from.username}) ${from.id} отправил фотографию`,
+                });
+                const messageUser = await ctx.reply('Ваша фотография передана Администратору');
+                setTimeout(() => ctx.deleteMessage(messageUser.message_id), 3000);
+            }
+
+            // Обработка голосового сообщения
+            if (message.voice) {
+                this.logger.info('Получено голосовое сообщение');
+                const voice = await ctx.telegram.getFile(message.voice.file_id);
+                await this.bot.telegram.sendVoice(config.adminId, voice.file_id, {
+                    caption: `Пользователь ${from.first_name} ${from.last_name} (@${from.username}) ${from.id} отправил голосовое сообщение`,
+                });
+                const messageUser = await ctx.reply('Ваше голосовое сообщение передано Администратору');
+                setTimeout(() => ctx.deleteMessage(messageUser.message_id), 3000);
+            }
+
+            // Обработка контакта
+            if (message.contact) {
+                this.logger.info('Получен контакт');
+                await this.bot.telegram.sendContact(
+                    config.adminId,
+                    message.contact.phone_number,
+                    message.contact.first_name,
+                    { last_name: message.contact.last_name }
+                );
+                const messageUser = await ctx.reply('Ваш контакт передан Администратору');
+                setTimeout(() => ctx.deleteMessage(messageUser.message_id), 3000);
+            }
         });
     }
 
@@ -133,6 +163,7 @@ class UserBot {
      */
     async cleanupOldRecords() {
         await CleanUpService.cleanupOldRecords();
+        await NotificationService.sendNotifications(this.bot);
         this.logger.info('Устаревшие записи очищены');
     }
 
@@ -161,7 +192,7 @@ class UserBot {
      */
     async stop() {
         await this.sendStopNotifications();
-        await this.bot.stop();
+        this.bot.stop();
         this.logger.info('Бот остановлен');
     }
 
@@ -171,10 +202,7 @@ class UserBot {
     async sendStopNotifications() {
         const users = await User.find();
         for (const user of users) {
-            await this.bot.telegram.sendMessage(
-                user.chatId,
-                'Бот временно недоступен.'
-            );
+            await this.bot.telegram.sendMessage(user.chatId, 'Бот временно недоступен.');
             this.logger.info(`Уведомление отправлено пользователю ${user.id}`);
         }
     }
